r9326/sofa-dev : FIX: compil
 //partial sync (15/20 changes)//

+++ framework/sofa/helper/fixed_array.h
@@ -308,7 +308,7 @@ public:
     void assign (const T& value)
     {
         //std::fill_n(begin(),size(),value);
-        for (size_type i=0; i<N; i++)
+        for (size_type i=0; i<(unsigned int)N; i++)
             elems[i] = value;
     }
